+++ CHANGELOG.md
@@ -8,3 +8,5 @@ Unreleased
 - [Add Deployment Pipeline](#1)
 - [Stub Out Jobs for Deployment Pipeline](#3)
 - [Add Unit Tests to Deployment](#5)
+- [Merge Local To Remote](#7)
+- [Add Readiness Check Endpoints](#9)